+++ frontend/src/pages/Resource/Resource.jsx
@@ -6,6 +6,7 @@ import {
   faCircleChevronDown,
   faCircleChevronRight,
 } from "@fortawesome/free-solid-svg-icons";
+import resourcesList from "../../data/resourcesList";
 
 const Resource = () => {
   const { resourceId } = useParams();
@@ -18,60 +19,6 @@ const Resource = () => {
       localStorage.setItem("checked", JSON.stringify(checked));
     }
   }, [checked]);
-
-  const resourcesList = [
-    {
-      name: "Web Development",
-      resourceId: "web-development",
-      resources: [
-        {
-          id: "Basic HTML & CSS",
-          headings: ["HTML and CSS", "Complete Css (advanced)"],
-          links: {
-            "HTML and CSS": "https://www.youtube.com/watch?v=a_iQb1lnAEQ",
-            "Complete Css (advanced)":
-              "https://www.youtube.com/watch?v=Edsxf_NBFrw",
-          },
-        },
-        {
-          id: "Styling",
-          headings: ["Bootstrap", "Tailwind CSS"],
-          links: {
-            Bootstrap: "https://www.youtube.com/watch?v=DUiYVJIVNcA",
-            "Tailwind CSS": "https://www.youtube.com/watch?v=MiR1dyW45Uw",
-          },
-        },
-        {
-          id: "Basic JS",
-          headings: ["JS with projects", "JQuery", "NodeJS"],
-          links: {
-            "JS with projects": "https://www.youtube.com/watch?v=hKB-YGF14SY",
-            JQuery: "https://www.youtube.com/watch?v=wjFfyqEutYE",
-            NodeJS: "https://www.youtube.com/watch?v=WjzLVU-aFXI",
-          },
-        },
-        {
-          id: "ReactJS",
-          headings: ["Basic to Intermediate", "Clone of Websites"],
-          links: {
-            "Basic to Intermediate":
-              "https://www.youtube.com/playlist?list=PLu0W_9lII9agx66oZnT6IyhcMIbUMNMdt",
-            "Clone of Websites":
-              "https://www.youtube.com/playlist?list=PL-J2q3Ga50oMQa1JdSJxYoZELwOJAXExP",
-          },
-        },
-        {
-          id: "TypeScript (TS)",
-          headings: ["Basics of TS", "ReactJS with TS"],
-          links: {
-            "Basics of TS": "https://www.youtube.com/watch?v=d56mG7DezGs",
-            "ReactJS with TS":
-              "https://www.youtube.com/watch?v=FJDVKeh7RJI&t=10s",
-          },
-        },
-      ],
-    },
-  ];
 
   const [currentResource, setCurrentResource] = useState({
     name: "",
